Added new test for changePlugin thunk
