+++ public/app/features/plugins/plugin_component.ts
@@ -5,7 +5,8 @@ import config from 'app/core/config';
 import coreModule from 'app/core/core_module';
 
 import { DataSourceApi, PanelEvents } from '@grafana/data';
-import { importPanelPlugin, importDataSourcePlugin, importAppPlugin } from './plugin_loader';
+import { importDataSourcePlugin, importAppPlugin } from './plugin_loader';
+import { importPanelPlugin } from './importPanelPlugin';
 import DatasourceSrv from './datasource_srv';
 import { GrafanaRootScope } from 'app/routes/GrafanaCtrl';
