+++ public/app/features/plugins/plugin_loader.ts
@@ -33,6 +33,7 @@ import * as emotion from '@emotion/css';
 import * as grafanaData from '@grafana/data';
 import * as grafanaUIraw from '@grafana/ui';
 import * as grafanaRuntime from '@grafana/runtime';
+import { GenericDataSourcePlugin } from '../datasources/settings/PluginSettings';
 
 // Help the 6.4 to 6.5 migration
 // The base classes were moved from @grafana/ui to @grafana/data
@@ -217,56 +218,4 @@ export function importAppPlugin(meta: grafanaData.PluginMeta): Promise<grafanaDa
     plugin.setComponentsFromLegacyExports(pluginExports);
     return plugin;
   });
-}
-
-import { getPanelPluginLoadError } from '../panel/components/PanelPluginError';
-import { GenericDataSourcePlugin } from '../datasources/settings/PluginSettings';
-
-interface PanelCache {
-  [key: string]: Promise<grafanaData.PanelPlugin>;
-}
-const panelCache: PanelCache = {};
-
-export function importPanelPlugin(id: string): Promise<grafanaData.PanelPlugin> {
-  const loaded = panelCache[id];
-  if (loaded) {
-    return loaded;
-  }
-
-  const meta = config.panels[id];
-
-  if (!meta) {
-    throw new Error(`Plugin ${id} not found`);
-  }
-
-  panelCache[id] = getPanelPlugin(meta);
-
-  return panelCache[id];
-}
-
-export function importPanelPluginFromMeta(meta: grafanaData.PanelPluginMeta): Promise<grafanaData.PanelPlugin> {
-  return getPanelPlugin(meta);
-}
-
-function getPanelPlugin(meta: grafanaData.PanelPluginMeta): Promise<grafanaData.PanelPlugin> {
-  return importPluginModule(meta.module)
-    .then((pluginExports) => {
-      if (pluginExports.plugin) {
-        return pluginExports.plugin as grafanaData.PanelPlugin;
-      } else if (pluginExports.PanelCtrl) {
-        const plugin = new grafanaData.PanelPlugin(null);
-        plugin.angularPanelCtrl = pluginExports.PanelCtrl;
-        return plugin;
-      }
-      throw new Error('missing export: plugin or PanelCtrl');
-    })
-    .then((plugin) => {
-      plugin.meta = meta;
-      return plugin;
-    })
-    .catch((err) => {
-      // TODO, maybe a different error plugin
-      console.warn('Error loading panel plugin: ' + meta.id, err);
-      return getPanelPluginLoadError(meta, err);
-    });
 }
